+++ Data/SBV/Control/Query.hs
@@ -12,7 +12,6 @@
 {-# LANGUAGE BangPatterns        #-}
 {-# LANGUAGE LambdaCase          #-}
 {-# LANGUAGE NamedFieldPuns      #-}
-{-# LANGUAGE Rank2Types          #-}
 {-# LANGUAGE ScopedTypeVariables #-}
 {-# LANGUAGE TupleSections       #-}
 {-# LANGUAGE ViewPatterns        #-}
@@ -40,21 +39,18 @@ import Control.Monad.IO.Class (MonadIO)
 import Data.IORef (readIORef)
 
 import qualified Data.Map.Strict as M
-import qualified Data.Sequence   as S
 import qualified Data.Text       as T
 import qualified Data.Foldable   as F
 
 
 import Data.Char      (toLower)
-import Data.List      (intercalate, nubBy, sortOn)
-import Data.Maybe     (listToMaybe, catMaybes, fromMaybe)
+import Data.List      (intercalate, nubBy)
+import Data.Maybe     (fromMaybe)
 import Data.Function  (on)
-import Data.Bifunctor (first)
-import Data.Foldable  (toList)
 
 import Data.SBV.Core.Data
 
-import Data.SBV.Core.Symbolic (MonadQuery(..), State(..), incrementInternalCounter, validationRequested, getSV, lookupInput, mustIgnoreVar)
+import Data.SBV.Core.Symbolic (MonadQuery(..), State(..), incrementInternalCounter, getSV)
 
 import Data.SBV.Utils.SExpr
 
@@ -66,17 +62,6 @@ import Data.SBV.Utils.PrettyNum (showNegativeNumber)
 
 -- | An Assignment of a model binding
 data Assignment = Assign SVal CV
-
--- | Remove the bars from model names; these are (mostly!) automatically inserted
-unBarModel :: SMTModel -> SMTModel
-unBarModel SMTModel {modelObjectives, modelBindings, modelAssocs, modelUIFuns}
-   = SMTModel { modelObjectives = ubf       <$> modelObjectives
-              , modelBindings   = (ubn <$>) <$> modelBindings
-              , modelAssocs     = ubf       <$> modelAssocs
-              , modelUIFuns     = ubf       <$> modelUIFuns
-              }
-   where ubf (n, a) = (unBar n, a)
-         ubn (NamedSymVar sv nm, a) = (NamedSymVar sv (T.pack (unBar (T.unpack nm))), a)
 
 -- Is this a string? If so, return it, otherwise fail in the Maybe monad.
 fromECon :: SExpr -> Maybe String
@@ -299,123 +284,6 @@ getParetoOptResults mbN      = do cfg <- getConfig
                                                        Sat    -> more
                                                        DSat{} -> more
                                                        Unk    -> more
-
--- | Generalization of 'Data.SBV.Control.getModel'
-getModel :: (MonadIO m, MonadQuery m) => m SMTModel
-getModel = getModelAtIndex Nothing
-
--- | Get a model stored at an index. This is likely very Z3 specific!
-getModelAtIndex :: (MonadIO m, MonadQuery m) => Maybe Int -> m SMTModel
-getModelAtIndex mbi = do
-    State{runMode} <- queryState
-    rm <- io $ readIORef runMode
-    case rm of
-      m@CodeGen     -> error $ "SBV.getModel: Model is not available in mode: " ++ show m
-      m@LambdaGen{} -> error $ "SBV.getModel: Model is not available in mode: " ++ show m
-      m@Concrete{}  -> error $ "SBV.getModel: Model is not available in mode: " ++ show m
-      SMTMode{}     -> do
-          cfg <- getConfig
-          uis <- getUIs
-
-          allModelInputs <- getTopLevelInputs
-          obsvs          <- getObservables
-
-          inputAssocs <- let grab (NamedSymVar sv nm) = let wrap !c = (sv, (nm, c)) in wrap <$> getValueCV mbi sv
-                         in mapM grab allModelInputs
-
-          let name     = fst . snd
-              removeSV = snd
-              prepare  = S.unstableSort . S.filter (not . mustIgnoreVar cfg . T.unpack . name)
-              assocs   = fmap removeSV (prepare inputAssocs) <> S.fromList (sortOn fst obsvs)
-
-          -- collect UIs, and UI functions if requested
-          let uiFuns = [ui | ui@(nm, (_, _, SBVType as)) <- uis, length as >  1, allSatTrackUFs cfg, not (mustIgnoreVar cfg nm)] -- functions have at least two things in their type!
-              uiRegs = [ui | ui@(nm, (_, _, SBVType as)) <- uis, length as == 1,                     not (mustIgnoreVar cfg nm)]
-
-          -- If there are uninterpreted functions, arrange so that z3's pretty-printer flattens things out
-          -- as cex's tend to get larger
-          unless (null uiFuns) $
-             let solverCaps = capabilities (solver cfg)
-             in case supportsFlattenedModels solverCaps of
-                  Nothing   -> return ()
-                  Just cmds -> mapM_ (send True) cmds
-
-          bindings <- let get i@(getSV -> sv) = case lookupInput fst sv inputAssocs of
-                                                  Just (_, (_, cv)) -> return (i, cv)
-                                                  Nothing           -> do cv <- getValueCV mbi sv
-                                                                          return (i, cv)
-
-                      in if validationRequested cfg
-                         then Just <$> mapM get allModelInputs
-                         else return Nothing
-
-          uiFunVals <- mapM (\ui@(nm, (c, _, t)) -> (\a -> (nm, (c, t, a))) <$> getUIFunCVAssoc mbi ui) uiFuns
-
-          uiVals    <- mapM (\ui@(nm, (_, _, _)) -> (nm,) <$> getUICVal mbi ui) uiRegs
-
-          return $ unBarModel $ SMTModel { modelObjectives = []
-                                         , modelBindings   = toList <$> bindings
-                                         , modelAssocs     = uiVals ++ toList (first T.unpack <$> assocs)
-                                         , modelUIFuns     = uiFunVals
-                                         }
-
--- | Just after a check-sat is issued, collect objective values. Used
--- internally only, not exposed to the user.
-getObjectiveValues :: forall m. (MonadIO m, MonadQuery m) => m [(String, GeneralizedCV)]
-getObjectiveValues = do let cmd = "(get-objectives)"
-
-                            bad = unexpected "getObjectiveValues" cmd "a list of objective values" Nothing
-
-                        r <- ask cmd
-
-                        si <- queryState >>= getSInfo
-
-                        inputs <- F.toList <$> getTopLevelInputs
-
-                        parse r bad $ \case EApp (ECon "objectives" : es) -> catMaybes <$> mapM (getObjValue si (bad r) inputs) es
-                                            _                             -> bad r Nothing
-
-  where -- | Parse an objective value out.
-        getObjValue :: SInfo -> (forall a. Maybe [String] -> m a) -> [NamedSymVar] -> SExpr -> m (Maybe (String, GeneralizedCV))
-        getObjValue si bailOut inputs expr =
-                case expr of
-                  EApp [_]          -> return Nothing            -- Happens when a soft-assertion has no associated group.
-                  EApp [ECon nm, v] -> locate nm v               -- Regular case
-                  _                 -> dontUnderstand (show expr)
-
-          where locate nm v = case listToMaybe [p | p@(NamedSymVar sv _) <- inputs, show sv == nm] of
-                                Nothing                          -> return Nothing -- Happens when the soft assertion has a group-id that's not one of the input names
-                                Just (NamedSymVar sv actualName) -> grab sv v >>= \val -> return $ Just (T.unpack actualName, val)
-
-                dontUnderstand s = bailOut $ Just [ "Unable to understand solver output."
-                                                  , "While trying to process: " ++ s
-                                                  ]
-
-                grab :: SV -> SExpr -> m GeneralizedCV
-                grab s topExpr
-                  | Just v <- recoverKindedValue si k topExpr = return $ RegularCV v
-                  | True                                      = ExtendedCV <$> cvt (simplify topExpr)
-                  where k = kindOf s
-
-                        -- Convert to an extended expression. Hopefully complete!
-                        cvt :: SExpr -> m ExtCV
-                        cvt (ECon "oo")                    = return $ Infinite  k
-                        cvt (ECon "epsilon")               = return $ Epsilon   k
-                        cvt (EApp [ECon "interval", x, y]) =          Interval  <$> cvt x <*> cvt y
-                        cvt (ENum    (i, _, _))            = return $ BoundedCV $ mkConstCV k i
-                        cvt (EReal   r)                    = return $ BoundedCV $ CV k $ CAlgReal r
-                        cvt (EFloat  f)                    = return $ BoundedCV $ CV k $ CFloat   f
-                        cvt (EDouble d)                    = return $ BoundedCV $ CV k $ CDouble  d
-                        cvt (EApp [ECon "+", x, y])        =          AddExtCV <$> cvt x <*> cvt y
-                        cvt (EApp [ECon "*", x, y])        =          MulExtCV <$> cvt x <*> cvt y
-                        -- Nothing else should show up, hopefully!
-                        cvt e = dontUnderstand (show e)
-
-                        -- drop the pesky to_real's that Z3 produces.. Cool but useless.
-                        simplify :: SExpr -> SExpr
-                        simplify (EApp [ECon "to_real", n]) = n
-                        simplify (EApp xs)                  = EApp (map simplify xs)
-                        simplify e                          = e
 
 -- | Generalization of 'Data.SBV.Control.checkSatAssuming'
 checkSatAssuming :: (MonadIO m, MonadQuery m) => [SBool] -> m CheckSatResult